Redefine ls define structure for windows
